refactor: hooks refactoring
